corrections:

+++ webserv_v1.0/srcs/exec/Response.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Response.cpp                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mmahfoud <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: sguillot <sguillot@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/22 22:12:59 by mmahfoud          #+#    #+#             */
-/*   Updated: 2024/10/02 00:43:34 by mmahfoud         ###   ########.fr       */
+/*   Updated: 2024/10/02 12:48:55 by sguillot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -291,7 +291,7 @@ std::map<std::string, std::string> Response::createEnvCgi()
 	std::map<std::string, std::string> env;
 	if (_client->getMethod() == "POST")
 	{
-		env["CONTENT_TYPE"] = _client->getContentType(); 
+		env["CONTENT_TYPE"] = _client->getContentType();
 		env["CONTENT_LENGTH"] = _client->getContentLength();
 	}
 	// SERVEUR_VAR
@@ -418,7 +418,7 @@ void Response::filePathFinder()
 
 std::string Response::ft_get()
 {
-	Server::log("Server's receive a GET request.", 1);
+	/* Server::log("Server's receive a GET request.", 1);
 	std::string content = readFileContent(this->_filePath);
 	std::string response = "";
 	Server::log("The file requested \"" + this->_filePath + "\" was found.", 1);
@@ -432,13 +432,60 @@ std::string Response::ft_get()
 	response += "Connection: close\r\n";
 	response += "Server: " + *this->_serverName.begin() + "\r\n\r\n";
 	response += content;
+	return (response); */
+
+	Server::log("Server's received a GET request.", 1);
+
+	if (access(this->_filePath.c_str(), F_OK) != 0)
+	{
+		std::string response = "HTTP/1.1 404 Not Found\r\n";
+		response += "Content-Type: text/html\r\n";
+		std::string content = "File not found";
+		std::ostringstream oss;
+		oss << content.size();
+		response += "Content-Length: " + oss.str() + "\r\n";
+		response += "Connection: close\r\n";
+		response += "Server: " + *this->_serverName.begin() + "\r\n\r\n";
+		response += content;
+		return (response);
+	}
+
+	std::string content = readFileContent(this->_filePath);
+
+	if (content.empty())
+	{
+		std::string response = "HTTP/1.1 500 Internal Server Error\r\n";
+		response += "Content-Type: text/html\r\n";
+		std::string error_content = "Error reading file content";
+		std::ostringstream oss;
+		oss << error_content.size();
+		response += "Content-Length: " + oss.str() + "\r\n";
+		response += "Connection: close\r\n";
+		response += "Server: " + *this->_serverName.begin() + "\r\n\r\n";
+		response += error_content;
+		return (response);
+	}
+
+	Server::log("The file requested \"" + this->_filePath + "\" was found.", 1);
+
+	std::string mimeType = getMimeType();
+	std::string response = firstHeader();
+	response += "Content-Type: " + mimeType + "\r\n";
+
+	std::ostringstream oss;
+	oss << content.size();
+	response += "Content-Length: " + oss.str() + "\r\n";
+	response += "Connection: close\r\n";
+	response += "Server: " + *this->_serverName.begin() + "\r\n\r\n";
+	response += content;
+
 	return (response);
 }
 
 /*response to a POST request*/
 std::string Response::ft_post()
 {
-	Server::log("Server's receive a POST request.", 1);
+/* 	Server::log("Server's receive a POST request.", 1);
 	std::string content = readFileContent(this->_filePath);
 	std::string response = "";
 
@@ -446,6 +493,53 @@ std::string Response::ft_post()
 
 	response = firstHeader();
 	response += "Content-Type: " + mimeType + "\r\n";
+	std::ostringstream oss;
+	oss << content.size();
+	response += "Content-Length: " + oss.str() + "\r\n";
+	response += "Connection: close\r\n";
+	response += "Server: " + *this->_serverName.begin() + "\r\n\r\n";
+	response += content;
+
+	return (response); */
+
+	Server::log("Server's received a POST request.", 1);
+
+	if (access(this->_filePath.c_str(), F_OK) != 0)
+	{
+		std::string response = "HTTP/1.1 404 Not Found\r\n";
+		response += "Content-Type: text/html\r\n";
+		std::string content = "File not found";
+		std::ostringstream oss;
+		oss << content.size();
+		response += "Content-Length: " + oss.str() + "\r\n";
+		response += "Connection: close\r\n";
+		response += "Server: " + *this->_serverName.begin() + "\r\n\r\n";
+		response += content;
+		return (response);
+	}
+
+	std::string content = readFileContent(this->_filePath);
+
+	if (content.empty())
+	{
+		std::string response = "HTTP/1.1 500 Internal Server Error\r\n";
+		response += "Content-Type: text/html\r\n";
+		std::string error_content = "Error reading file content";
+		std::ostringstream oss;
+		oss << error_content.size();
+		response += "Content-Length: " + oss.str() + "\r\n";
+		response += "Connection: close\r\n";
+		response += "Server: " + *this->_serverName.begin() + "\r\n\r\n";
+		response += error_content;
+		return (response);
+	}
+
+	std::string mimeType = getMimeType();
+	std::string response = firstHeader();
+
+	response += "HTTP/1.1 201 Created\r\n";
+	response += "Content-Type: " + mimeType + "\r\n";
+
 	std::ostringstream oss;
 	oss << content.size();
 	response += "Content-Length: " + oss.str() + "\r\n";
@@ -476,6 +570,21 @@ std::string Response::ft_delete()
 		return (response);
 	}
 
+	if (access(getFilePath().c_str(), W_OK) != 0)
+	{
+		std::string response = "HTTP/1.1 403 Forbidden\r\n";
+		response += "Content-Type: text/html\r\n";
+		std::string content = "Permission denied";
+		std::ostringstream oss;
+		oss << content.size();
+		response += "Content-Length: " + oss.str() + "\r\n";
+		response += "Connection: close\r\n";
+		response += "Server: " + *this->_serverName.begin() + "\r\n\r\n";
+		response += content;
+
+		return (response);
+	}
+
 	if (unlink(getFilePath().c_str()) == 0)
 	{
 		std::string response = "HTTP/1.1 204 No Content\r\n";
@@ -485,8 +594,20 @@ std::string Response::ft_delete()
 		return (response);
 	}
 	else
+	{
 		Server::log("Server failed to respond at the DELETE request.", 2);
-	return ("");
+		std::string response = "HTTP/1.1 500 Internal Server Error\r\n";
+		response += "Content-Type: text/html\r\n";
+		std::string content = "Internal server error";
+		std::ostringstream oss;
+		oss << content.size();
+		response += "Content-Length: " + oss.str() + "\r\n";
+		response += "Connection: close\r\n";
+		response += "Server: " + *this->_serverName.begin() + "\r\n\r\n";
+		response += content;
+
+		return (response);
+	}
 }
 
 /*response to a bad request*/
